requirements.txt
annotated-types==0.7.0
anyio==4.11.0
azure-core==1.35.1
azure-storage-blob==12.26.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
fastapi==0.118.0
h11==0.16.0
idna==3.10
isodate==0.7.2
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
python-multipart==0.0.20
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
